style(hydro_lang): use where clauses to increase readability (#1827)
A minor issue in the grand scheme of things, but I caught myself
squinting when trying to tease apart different trait implementations
specially when following the "implementations" popup navigation. In
addition, some of the generic variable naming was not consistent and a
bit confusing to the uninitiated. For example, `Order` was a generic
variable, but `TotalOrder` and `NoOrder` were a type. Not a big deal
when you already know the API, but just adds friction someone new to the
project.

Here are some of the items that guided the PR:
* Single capital letter for generic variables following a loose
standard: T for types, O for ordering, B for boundedness, L for
location, P for process or external process, C for cluster, etc.
* Use `where` clauses to keep the head of traits and functions lean and
more easily comparable.