feat(hydro_deploy): update how progress is displayed, fix #1415 (#1845)
Fix #1415

Changes the display of progress to be a bit more intuitive.  
Now the format is `finished / started_or_finished / total`  
Example progress:
1. `0/1/5`
2. `1/1/5`
3. `1/5/5`
4. `3/5/5`
5. `5/5/5` (complete)

This is more natural as all numbers increase as progress increases, and
all numbers are 100% when complete.

A question mark `?` will be displayed as the total if no total is given
or the given total is invalid.

---

### Previously...

Previously displayed as: `finished / started_but_not_finished / total`
Example progress:
1. `0/1/5`
2. `1/0/5`
3. `1/4/5`
4. `3/2/5`
5. `5/0/5` (complete)

This is confusing as the `started_but_not_finished` column goes up and
down as progress increases, and ends at zero.