refactor(hydro_lang)!: set up prelude, move collections under `live_collections` module (#2067)
Bigger refactor to clean up the top-level namespace for `hydro_lang`.
First, we move all the definitions for different live collections into
the `live_collections` module.

We also rename the `unsafety` module to `nondet`.

Then, instead of having a bunch of types exported directly from
`hydro_lang`, we instead create a `prelude` module that exports them.
This reduces the pollution of the namespace when importing from
`hydro_lang`.