feat(hydro_lang): ensure sim logs show up in captured stdout/stderr (#2180)
When running a sim test with `cargo test`, the stdout / stderr are
captured. But this capturing mechanism only works for direct calls to
`println!` and `eprintln!`, so the previous logic to use
`std::io::eprintln` does not work (and logs are displayed out of order).

To fix this, we change the logger to be a tri-state `LogKind`. If it is
`LogKind::Stderr`, we directly invoke `eprintln`. We also pass down
function pointers for `println` and `eprintln` to the staged code, where
we shadow `std::{println, eprintln, debug}` with macros that call the
function pointer. This ensures that any logging statements in quoted
code show up in the right place next to simulator logs.