feat(hydro_lang): avoid redundant simulations for ordering within unordered batches (#2183)
Previously, we would select elements for the next released batched in
non-deterministic order. This is unnecessary, because the batch itself
has `NoOrder` guarantees and so the permutations within the batch are
redundant.

Reduces instance count from 192 to 75 for
`sequence_payloads_sequences_all`. To avoid regressions,
`.sim().exhaustive()` now returns the number of simulation instances,
which we can verify against a combinatorial formula.